Add audioapp sample which shows the embedded album art of all songs on the device
Summary: Using a RecyclerView with 3 columns and square items, a grid of pictures is displayed. One item represents one MediaStore audio entry on the device which is classified as music. A SVG drawable is programmatically set as a fallback failure image. This app is supposed to showcase the new LocalAudioThumbnailProducer which handles all 'audio/' MimeTypes and tries to extract an embedded picture. Accessing the database of available songs requires the READ_EXTERNAL_STORAGE permission which is requested at runtime if not given.